move FakeAndroidKeystore to its own module and setup publishing of it to Sonatype
