+++ encrypted-storage/build.gradle
@@ -58,6 +58,7 @@ dependencies {
     implementation 'androidx.security:security-crypto:1.0.0-rc02'
 
     testImplementation 'junit:junit:4.12'
+    testImplementation project(':fake-keystore')
     testImplementation "org.jetbrains.kotlin:kotlin-test-junit:$kotlin_version"
     testImplementation "org.robolectric:robolectric:4.3.1"
     testImplementation 'androidx.test:core:1.2.0'
