+++ encrypted-storage/gradle.properties
@@ -1,0 +1,6 @@
+POM_ARTIFACT_ID=encrypted-storage
+POM_NAME=encrypted-storage
+POM_DESCRIPTION=An Android library for creating encrypted Files and SharedPreferences
+
+POM_LICENCE_NAME=GNU General Public License, Version 3.0
+POM_LICENCE_URL=https://www.gnu.org/licenses/gpl-3.0.html
