+++ fake-keystore/gradle.properties
@@ -1,0 +1,6 @@
+POM_ARTIFACT_ID=fake-keystore
+POM_NAME=fake-keystore
+POM_DESCRIPTION=A Fake Android Keystore for testing
+
+POM_LICENCE_NAME=Apache License, Version 2.0
+POM_LICENCE_URL=http://www.apache.org/licenses/LICENSE-2.0
