+++ gradle.properties
@@ -21,21 +21,16 @@ android.enableJetifier=true
 kotlin.code.style=official
 
 GROUP=io.matthewnelson.encrypted-storage
-POM_ARTIFACT_ID=encrypted-storage
 VERSION_NAME=1.0.3-SNAPSHOT
 VERSION_CODE=103
 
-POM_NAME=encrypted-storage
-POM_DESCRIPTION=An Android library for creating encrypted Files and SharedPreferences
 POM_INCEPTION_YEAR=2020
 
 POM_URL=https://github.com/05nelsonm/encrypted-storage
 POM_SCM_URL=https://github.com/05nelsonm/encrypted-storage
 POM_SCM_CONNECTION=scm:git:git://github.com/05nelsonm/encrypted-storage.git
 POM_SCM_DEV_CONNECTION=scm:git:ssh://git@github.com/05nelsonm/encrypted-storage.git
-
-POM_LICENCE_NAME=GNU General Public License, Version 3.0
-POM_LICENCE_URL=https://www.gnu.org/licenses/gpl-3.0.html
+        
 POM_LICENCE_DIST=repo
 
 POM_DEVELOPER_ID=05nelsonm
