+++ settings.gradle
@@ -1,3 +1,4 @@
+include ':fake-keystore'
 include ':encrypted-storage'
 include ':demo'
 rootProject.name = "Encrypted Storage"
